Rename programs and add floating-point number support to Reverse Polish Notation

+++ Assignment2/Assignment2b/main.cpp
@@ -14,13 +14,13 @@ int main()
 	const size_t stack_size = (input_size + 1) / 2;
 	if (input[input_size - 1] != ' ') input = input + ' ';
 
-	int* stack = new int[stack_size];
-	int* header = stack - 1;
+	double* stack = new double[stack_size];
+	double* header = stack - 1;
 	size_t  index = 0;
 
 	while (index < input_size)
 	{
-		int result;
+		double result;
 		const size_t start = index;
 
 		while(input[index] != ' ') index++;
@@ -35,8 +35,8 @@ int main()
 
 		if (token.size() == 1 && (oper_add || oper_sub || oper_prod || oper_div))
 		{
-			int num1 = *(header--);
-			int num2 = *(header--);
+			double num1 = *(header--);
+			double num2 = *(header--);
 			is_number = false;
 			if (oper_add) result = num2 + num1;
 			else if(oper_sub) result = num2 - num1;
@@ -44,8 +44,7 @@ int main()
 			else result = num2 / num1;
 		}
 
-		if(is_number) result = stoi(token);
-
+		if(is_number) result = stod(token);
                	*(++header) = result;
 		index++;
 	}
